+++ CaffParser/caff.h
@@ -5,19 +5,19 @@
 class Caff
 {
 private:
-    uint headerSize;
-    uint numAnim;
+    ull headerSize;
+    ull numAnim;
     byte[] year;
     byte month;
     byte day;
     byte hour;
     byte minute;
-    uint creatorLen;
+    ull creatorLen;
     string creator;
-    uint duration;
+    ull duration;
     vector<CIFF> ciffs;
 
-    Caff(uint hs, uint na, byte[]& ye, byte mo, byte da, byte ho, byte mi, uint crlen, const string& cr, uint dur, const vector<CIFF>& ci):
+    Caff(ull hs, ull na, byte[]& ye, byte mo, byte da, byte ho, byte mi, ull crlen, const string& cr, ull dur, const vector<CIFF>& ci):
         headerSize(hs), numAnim(na), year(ye), month(mo), day(da), hour(ho), minute(mi), creatorLen(crlen), creator(cr), duration(dur), ciffs(ci) {};
 
     Caff(){};
@@ -28,18 +28,18 @@ public:
     static const int CREDIT_STAGES = 7;
     static const int ANIMATION_STAGES = 2;
 
-    inline uint getHeaderSize() { return headerSize; }
-    inline uint getNumAnim() { return numAnim; }
+    inline ull getHeaderSize() { return headerSize; }
+    inline ull getNumAnim() { return numAnim; }
     inline byte[] getYear() { return year; }
     inline byte getMonth() { return month; }
     inline byte getDay() { return day; }
     inline byte getHour() { return hour; }
     inline byte getMinute() { return minute; }
-    inline uint getCreatorLen() { return creatorLen; }
+    inline ull getCreatorLen() { return creatorLen; }
     inline string getCreator() { return creator; }
-    inline uint getDuration() { return duration; }
+    inline ull getDuration() { return duration; }
 
-    CIFF getCIFF(uint);
+    CIFF getCIFF(ull);
     static Caff parse(istream&, int , int);
 };
 */
